requirements.txt
anyio==3.5.0
asgiref==3.4.1
blas==1.0
bzip2==1.0.8
ca-certificates==2021.10.26
certifi==2021.10.8
click==7.1.2
fastapi==0.65.0
ffmpeg==4.3
freetype==2.11.0
gettext==0.21.0
giflib==5.2.1
gmp==6.2.1
gnutls==3.6.15
h11==0.12.0
icu==58.2
idna==3.3
intel-openmp==2021.4.0
jpeg==9b
lame==3.100
lcms2==2.12
libcxx==12.0.0
libffi==3.3
libiconv==1.16
libidn2==2.3.2
libpng==1.6.37
libtasn1==4.16.0
libtiff==4.2.0
libunistring==0.9.10
libuv==1.40.0
libwebp-base==1.2.0
libwebp==1.2.0
libxml2==2.9.12
llvm-openmp==12.0.0
lz4-c==1.9.3
mkl-service==2.4.0
mkl==2021.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
ncurses==6.3
nettle==3.7.3
ninja==1.10.2
numpy-base==1.21.2
numpy==1.21.2
olefile==0.46
openh264==2.1.1
openssl==1.1.1m
pillow==9.0.0
pip==21.2.4
pydantic==1.9.0
python-multipart==0.0.5
python==3.8.12
pytorch==1.8.1
readline==8.1.2
setuptools==58.0.4
six==1.16.0
sniffio==1.2.0
sqlite==3.37.0
starlette==0.14.2
tk==8.6.11
torchaudio==0.8.1
torchvision==0.9.1
typing_extensions==3.10.0.2
uvicorn==0.13.3
wheel==0.37.1
xz==5.2.5
zlib==1.2.11
zstd==1.4.9
